moved magical girl type implementations to a dedicated folder
