+++ Assets/Scripts/MagicalGirls/MagicalGirlTypes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ee80bd4041d34bd4f88328b4d2681212
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
