+++ Assets/Scripts/MagicalGirls/MagicalGirlTypes/AngryStatePlaceholder.cs.meta
