+++ Assets/Scripts/MagicalGirls/MagicalGirlTypes/HappyStatePlaceholder.cs.meta
